refactor: update features with eslint rules
