+++ src/features/category/ui/Categories/index.ts
@@ -1,0 +1,1 @@
+export { Categories } from "./Categories";
