+++ src/features/pagination/index.ts
@@ -1,3 +1,1 @@
-import Pagination from "./ui/Pagination/Pagination";
-
-export { Pagination };
+export { Pagination } from "./ui";
