+++ src/features/pagination/model/types.ts
@@ -2,6 +2,6 @@ export interface IPaginationProps {
   totalPages: number;
   handlePreviousPage: () => void;
   handleNextPage: () => void;
-  handlePageClick: (page: number) => void;
+  handlePageClick: (_page: number) => void;
   currentPage: number;
 }
