+++ src/features/pagination/ui/Pagination/index.ts
@@ -1,0 +1,1 @@
+export { Pagination } from "./Pagination";
