+++ src/features/search/index.ts
@@ -1,3 +1,1 @@
-import Search from "./ui/Search/Search";
-
-export { Search };
+export { Search } from "./ui";
