+++ src/features/search/ui/Search/index.ts
@@ -1,0 +1,1 @@
+export { Search } from "./Search";
