+++ src/features/slider/ui/Slider/index.ts
@@ -1,0 +1,1 @@
+export { Slider } from "./Slider";
