+++ src/features/theme/ui/ThemeButton/index.ts
@@ -1,0 +1,1 @@
+export { ThemeButton } from "./ThemeButton";
