+++ src/features/theme/ui/index.ts
@@ -1,0 +1,1 @@
+export { ThemeButton } from "./ThemeButton";
